stqcf-backend/requirements/base.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-environ==0.11.2
celery==5.3.0
cassandra-driver==3.28.0
django-cassandra-engine==1.8.0
mysqlclient==2.2.0
pycryptodome==3.19.0
grpcio==1.59.0
protobuf==4.24.0
python-jose==3.3.0
gunicorn==21.2.0
pytest==7.4.0
black==23.9.0
flake8==6.1.0
django-extensions==3.2.3
django-debug-toolbar==4.2.0
djangorestframework-simplejwt==5.3.0
docker==7.0.0 